requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.0
datasets>=2.18.0
accelerate>=0.28.0
bitsandbytes>=0.43.0

# Orchestration
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
llama-index>=0.10.0

# AWS
boto3>=1.34.0
python-dotenv>=1.0.0

# API & Frontend
fastapi>=0.111.0
uvicorn>=0.29.0
streamlit>=1.35.0
pydantic>=2.7.0

# MLflow
mlflow>=2.13.0

# Data
pandas>=2.2.0
numpy>=1.26.0

# Testing
pytest>=8.2.0
httpx>=0.27.0
